Create a workflow to apply infrastructure changes
Along with the "preview" workflow that runs when a PR is opened,
add an "update" workflow that will apply the infrastructure updates
when a PR is merged to the main branch.

Note that both worfklows assume the same Pulumi IAM role while
we're in demo mode. If we adopted an IaC tool like this, we'd
want separate IAM roles: one to preview (w/ read-only access
that can be assumed by any branch in the repo) and one for
the updates (with "create" access that can only be assumed by
the main branch)