Use pip-tools to generate "locked" requirements files
Although we've been using PDM to manage dependencies and lockfiles,
that process isn't working well in CI when combined with the Pulumi
setup (the problem is related to how each of those tools works
with Python virtual environments). Rather than fight the tools,
let's go back to the tried and true requirements.txt (with the
addition of pip-tools, so we can get the documentation that provides
insight into the dependency tree).