Updated PR for switching to src layout (#31)
* Upgrade to ye olde comprehensive Python .gitignore

* Switch to pdm + src layout

Re-organize the repo to follow the Python layout and tooling as
described here: https://reichlab.atlassian.net/wiki/spaces/~7120203df9acbaa41944839969b0fa9d27fea5/pages/1835009/Python+setup+notes+on+new+laptop#Creating-a-new-project

* Consolidate and update README files

This is a first pass. We need more detail, especially wrt
project setup, but let's wait until we're done re-arranging things.

* Use pip-tools to generate "locked" requirements files

Although we've been using PDM to manage dependencies and lockfiles,
that process isn't working well in CI when combined with the Pulumi
setup (the problem is related to how each of those tools works
with Python virtual environments). Rather than fight the tools,
let's go back to the tried and true requirements.txt (with the
addition of pip-tools, so we can get the documentation that provides
insight into the dependency tree).

* Update workflows to revert PDM changes and respect new src layout

* Remove pdm.lock

* README tweaks

* fix rebase issues

* Add .python-version file

The setup-python GitHub action will use this file to set the
Python version (so you don't need to do so explicitly). Also,
this file is respected by pyenv and many IDEs, so handy to have it.

* Remove working directory default

* Remove the PDM lockfile