Update Pulumi workflows to use different AWS roles
Now that we're out of demo mode, apply the principle of least privilege to our Pulumi workflows.
The "preview" workflow should only have read access to AWS resources, especially since it runs
against feature branch code. The role with write access is reserved for the update action,
which runs when a PR is merged to the repo's main branch.