Add function to create the model-output transform lambda
Create the lamba that will be triggered when new model-output files
are pushed to a hub's S3 bucket. This definition points to a lambda
package on S3 (rather than defining the function code inline). Another
repo will be responsible for creating the the lambda package and deploying
to the s3 bucket.