Create a placeholder lambda package on S3
Annoyingly, the "create lambda" function will fail if it's pointing
to a lambda code package that doesn't yet exist on S3. It creates a
chicken-and-egg problem for us, since we haven't deployed the transform
function's code to S3 yet. Might be overkill, but this changeset
creates a placeholder .zip to use as lambda code package until
we have the official deployment pipeline up and running (in the
hubverse-transform repo)