Add an explicit id for CORS rules
Issuing a Pulumi refresh command from my local machine indicates that the
state config for our Hubverse S3 bucket CORS rules is out of sync with actual
AWS state of these rules.

The only difference I could find between Pulumi's stack data and the config
as viewed from AWS is the id. On AWS, the CORS JSON doesn't have an id key,
and in the Pulumi stack, there is an id field set to "". I'm not sure how
the id field got there but, let's set an explicit id and hope we don't see
this kind of diff going forward.