Try bumping the max number of open Dependabot requests
Dependabot continues to submit PRs that fail CI checks because pip
cannot resolve the suggested dependency versions. This week, the
Dependabot PR tried to update 5 packages.

By default, Dependabot will allow 5 open PRs at a time, so this changeset
updates that number to see if it will cause a larger number of pacakge
updates at once (with the idea that including all package updates will
help the dependencies resolve)

Not sure this will do anything, since we group the PRs, but it also won't
hurt anything.